+++ pyproject.toml
@@ -1,5 +1,5 @@
 [tool.poetry]
-name = "pantry"
+name = "data-file-merge"
 version = "0.1.0"
 description = ""
 authors = ["ServerlessSam"]
