+++ tests/test_files_directory/nested_directory/build_test_merged_file.json
@@ -1,7 +1,3 @@
 {
-    "Hello": "There",
-    "UhOh": "This",
-    "OneIs": "Nested",
-    "UhOh2": "This",
-    "OneIs2": "NestedAlso"
+    "Hello": "There"
 }
